Refactor example and update readme
